add a createReviewSubsection() method
